requirements.txt
Django==1.10.4
requests==2.12.4
sorl-thumbnail==12.3
bibtexparser==0.6.2
pillow==4.0.0
django-extended-choices==1.1.1
django-bootstrap3==8.1.0
django-post-office==2.0.8

